feat: admin-api: add endpoint to update host cost
Fixes https://github.com/codius/codiusd/issues/42

Adds the `POST /config` endpoint to the admin API. Currently the only property that can be updated is `hostCostPerMonth`.

The profit computation for `aggregate_earnings` is updated to use ilp-stream's `totalReceived`, so it should be accurate in the case of errors or rate changes.